refactor: move ImageLib into Librelancer.Base
