+++ src/Editor/LibreLancer.ImUI/LibreLancer.ImUI.csproj
@@ -20,7 +20,6 @@
   <ItemGroup>
     <ProjectReference Include="..\..\LibreLancer.Base\LibreLancer.Base.csproj" />
     <ProjectReference Include="..\..\LibreLancer.Data\LibreLancer.Data.csproj" />
-    <ProjectReference Include="..\..\LibreLancer.ImageLib\LibreLancer.ImageLib.csproj" />
     <ProjectReference Include="..\..\LibreLancer.Media\LibreLancer.Media.csproj" />
   </ItemGroup>
