+++ src/Editor/lleditscript/lleditscript.csproj
@@ -9,7 +9,6 @@
     <ItemGroup>
       <ProjectReference Include="..\..\LibreLancer.Base\LibreLancer.Base.csproj" />
       <ProjectReference Include="..\..\LibreLancer.Data\LibreLancer.Data.csproj" />
-      <ProjectReference Include="..\..\LibreLancer.ImageLib\LibreLancer.ImageLib.csproj" />
       <ProjectReference Include="..\..\LibreLancer.Media\LibreLancer.Media.csproj" />
       <ProjectReference Include="..\LibreLancer.ContentEdit\LibreLancer.ContentEdit.csproj" />
     </ItemGroup>
