+++ src/LibreLancer.Base/ImageLib/Image.cs
