--- src/LibreLancer.ImageLib/LibreLancer.ImageLib.csproj
@@ -1,29 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <ImportGroup>
-    <Import Project="../CommonVersion.props"/>
-  </ImportGroup>
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-  </PropertyGroup>
-
-  <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-
-  <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\LibreLancer.Base\LibreLancer.Base.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="StbImageSharp" Version="2.22.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Properties" />
-  </ItemGroup>
-
-</Project>
