+++ src/LibreLancer/LibreLancer.csproj
@@ -25,7 +25,6 @@
     <ProjectReference Include="..\LibreLancer.Data\LibreLancer.Data.csproj" />
     <ProjectReference Include="..\LibreLancer.Data.Generator\LibreLancer.Data.Generator.csproj" ReferenceOutputAssembly="false" OutputItemType="Analyzer" />
     <ProjectReference Include="..\LibreLancer.Entities\LibreLancer.Entities.csproj" />
-    <ProjectReference Include="..\LibreLancer.ImageLib\LibreLancer.ImageLib.csproj" />
     <ProjectReference Include="..\LibreLancer.Media\LibreLancer.Media.csproj" />
     <ProjectReference Include="..\LibreLancer.Physics\LibreLancer.Physics.csproj" />
     <ProjectReference Include="..\LibreLancer.Thorn\LibreLancer.Thorn.csproj" />
